orders/src/events/listener/product-updated-listener.ts
import { Message } from 'node-nats-streaming';
import { Subjects, Listener, ProductUpdatedEvent } from '@picktronics/common';
import { Product } from '../../models/product';
import { queueGroupName } from './queue-group-name';

export class ProductUpdatedListener extends Listener<ProductUpdatedEvent> {
  subject: Subjects.ProductUpdated = Subjects.ProductUpdated;
  queueGroupName = queueGroupName;

  async onMessage(data: ProductUpdatedEvent['data'], msg: Message) {
    const product = await Product.findByEvent(data);

    if (!product) {
      return new Error('Product not found');
    }

    const { title, price } = data;
    product.set({ title, price });
    await product.save();

    msg.ack();
  }
}

// Might need to check later because there is a bug on repetitive ack on update
// listener event
